Navigator being implemented

+++ idp.h
@@ -24,7 +24,7 @@ public:
 
 };
 
-inline int diff_ms(timeval t1, timeval t2){
+inline double diff_ms(timeval t1, timeval t2){
     return ((t1.tv_sec - t2.tv_sec) * 1000000 + (t1.tv_usec - t2.tv_usec))/1000;
 }
